refactor(cli): reorganize — util files to util/, tests to test/, rename json-schema-to-yargs → typebox-to-yargs
