+++ packages/cli/src/cli.ts
@@ -3,7 +3,7 @@ import { buildAuthCommand } from './commands/auth-command';
 import { buildDataCommand } from './commands/data-command';
 import { buildStartCommand } from './commands/start-command';
 import { buildWorkspaceCommand } from './commands/workspace-command';
-import { resolveEpicenterHome } from './paths';
+import { resolveEpicenterHome } from './util/paths';
 
 /**
  * Create the Epicenter CLI instance.
