+++ packages/cli/src/util/format-output.ts
