+++ packages/cli/src/util/paths.ts
